--- src/react/dry-run-context.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const DryRunContext = React.createContext(false);
-
-export default DryRunContext;
